[CI] Move manual-only nightly tests out of test/registered/
Move test_qwen3_235b.py, test_deepseek_v31.py, and test_glm_46_fp8.py
from test/registered/8-gpu-models/ to test/manual/. These are
manual-only nightly tests that don't belong in test/registered/
(which requires a CI registry entry). Having them there breaks
collect_tests() with: ValueError: No CI registry found.

Also remove the register_cuda_ci() calls from deepseek_v31 and
glm_46_fp8 since they are manual tests.
